Conflict resolution data class moving to basemodel
