+++ notifications/src/main/kotlin/com/amazon/opendistroforelasticsearch/commons/notifications/action/GetFeatureChannelListRequest.kt
@@ -30,6 +30,7 @@ import org.elasticsearch.common.xcontent.XContentBuilder
 import org.elasticsearch.common.xcontent.XContentParser
 import org.elasticsearch.common.xcontent.XContentParserUtils
 import java.io.IOException
+import kotlin.jvm.Throws
 
 /**
  * This request is plugin-only call. i.e. REST interface is not exposed.
