+++ notifications/src/main/kotlin/com/amazon/opendistroforelasticsearch/commons/notifications/model/FeatureChannel.kt
@@ -26,6 +26,7 @@ import org.elasticsearch.common.xcontent.XContentBuilder
 import org.elasticsearch.common.xcontent.XContentParser
 import org.elasticsearch.common.xcontent.XContentParserUtils
 import java.io.IOException
+import kotlin.jvm.Throws
 
 /**
  * Data class representing Notification config for exposed for other plugins.
