+++ notifications/src/main/kotlin/com/amazon/opendistroforelasticsearch/commons/notifications/model/channel/Chime.kt
@@ -16,8 +16,8 @@
 package com.amazon.opendistroforelasticsearch.commons.notifications.model.channel
 
 import com.amazon.opendistroforelasticsearch.commons.notifications.model.NotificationConfigType
-import com.amazon.opendistroforelasticsearch.notifications.util.logger
-import com.amazon.opendistroforelasticsearch.notifications.util.validateUrl
+import com.amazon.opendistroforelasticsearch.commons.utils.logger
+import com.amazon.opendistroforelasticsearch.commons.utils.validateUrl
 import org.elasticsearch.common.Strings
 import org.elasticsearch.common.io.stream.StreamInput
 import org.elasticsearch.common.io.stream.StreamOutput
